Using 64 bit ints in andersson tree

+++ include/c4/klib/andtree.h
@@ -1,11 +1,12 @@
 #ifndef _C4_ANDTREE_H
 #define _C4_ANDTREE_H 1
 #include <stdbool.h>
+#include <stdint.h>
 
 typedef struct c4_anode c4_anode_t;
 typedef struct c4_atree c4_atree_t;
 
-typedef int (*c4_atree_get_key_t)(void *b);
+typedef int64_t (*c4_atree_get_key_t)(void *b);
 
 struct c4_anode {
 	c4_anode_t *parent;
@@ -23,7 +24,7 @@ struct c4_atree {
 c4_anode_t *atree_insert( c4_atree_t *tree, void *data );
 
 void atree_remove( c4_atree_t *tree, c4_anode_t *node );
-void *atree_remove_key( c4_atree_t *tree, int key );
+void *atree_remove_key( c4_atree_t *tree, int64_t key );
 void *atree_remove_data( c4_atree_t *tree, void *data );
 
 c4_atree_t *atree_init( c4_atree_t *tree,
@@ -32,9 +33,9 @@ c4_atree_t *atree_deinit( c4_atree_t *tree );
 void atree_print( c4_atree_t *tree );
 bool atree_check( c4_atree_t *tree );
 
-c4_anode_t *atree_find_at_least( c4_atree_t *tree, int key );
+c4_anode_t *atree_find_at_least( c4_atree_t *tree, int64_t key );
 c4_anode_t *atree_find_data( c4_atree_t *tree, void *data );
-c4_anode_t *atree_find_key( c4_atree_t *tree, int key );
+c4_anode_t *atree_find_key( c4_atree_t *tree, int64_t key );
 
 c4_anode_t *atree_start(c4_atree_t *tree);
 c4_anode_t *atree_end(c4_atree_t *tree);
